cache command should not prune gems for other platforms

+++ lib/bundler/runtime.rb
@@ -99,6 +99,7 @@ module Bundler
       Bundler.ui.info "Removing outdated .gem files from vendor/cache"
       cache_path.children.each do |gem_path|
         cached_spec = Gem::Format.from_file_by_path(gem_path).spec
+        next unless cached_spec.platform =~ Gem::Platform.local
         unless specs.any?{|s| s.full_name == cached_spec.full_name }
           Bundler.ui.info "  * #{File.basename(gem_path)}"
           gem_path.rmtree
